feat: callback-driven order tracking, Redis bridge, xuntou refactor
- Add xt_status/xt_error_msg/xt_traded_volume/xt_traded_price/order_price_type to pending_orders
- Refactor XunTou SDK into adapters/xuntou/ (base_service, order_service, query_service, connection_manager)
- Callback-driven order status updates via onOrderEvent/onTradeEvent
- Replace polling with manual reconciliation endpoint
- Add price_service (Redis tick cache), remote_bus (Redis bridge)
- Add market health API, fund inflow management, order confirmation flow
- LD_LIBRARY_PATH isolation between CTP and XunTou processes
- Frontend: order-list.html, order-confirm.html, fund-inflow.html, health.html
