+++ archive/tests_and_demos/test_real_order.py
@@ -9,8 +9,8 @@ os.environ['LD_LIBRARY_PATH'] = '/home/ubuntu/xt_sdk:' + os.environ.get('LD_LIBR
 from decimal import Decimal
 from database.dao import DatabaseDAO
 from database.models import PendingOrder, OrderStatus
-from adapters.xt_connection_manager import get_xt_connection_manager
-from adapters.xt_order_service import OrderRequest, OrderDirection, OrderPriceType
+from adapters.xuntou import get_xt_connection_manager
+from adapters.xuntou import OrderRequest, OrderDirection, OrderPriceType
 import uuid
 import time
