--- reminder.txt
@@ -1,3 +1,0 @@
-1.规则可能不止一些技术指标这么简单
-2.回测系统如何做是个大问题
-3.设计一个user friendly的前端界面也很重要
