--- tests/__init__.py
@@ -1,1 +1,0 @@
-"""测试代码"""
